adjust signer/authority requirements
- creating an auth account is now permissionless (not requiring the
  target owner of the account to sign in order to create it)
- the `airdrop_close` instruction now uses a hardcoded address for the
  authority (set to the jet eng multisig) instead of relying on the
  original signer that setup the airdrop.

- minor lint updates

+++ programs/staking/src/state.rs
@@ -340,7 +340,7 @@ impl FullAmount {
         assert!((share_amount > 0 && token_amount > 0) || (share_amount == 0 && token_amount == 0));
 
         let share_amount = share_amount as u64;
-        let token_amount = token_amount as u64;
+        let token_amount = token_amount;
 
         Self {
             token_amount,
